Baseline version for four Stokes inversion codes
This is a compilation of various spectropolarimetric inversion codes
that were covered in the 2022 NCAR/NSO HAO Summer Spectropolarimetry
School. Where available, github repos are used as code sources.
SIR and DeSIRe are two exceptions with newer (SIR) or unavailable
(DeSIRe) sources.

An MPI container is used as a basis for the inversion containers
so that, where appropriate, they can be scaled and parallelized
using mpi4py/openmpi.

See README.md for attributions

Curt Dodds 2022-08-23